+++ armor.h
@@ -6,7 +6,7 @@
 using namespace std;
 
 void text(string szText,short X,short Y,WORD color);
-void num(int num,int X,int Y,WORD color);
+void num(int num, short X, short Y,WORD color);
 
 class Armor : public Item{
 public:
