+++ player.h
@@ -14,7 +14,7 @@
 
 string rotate(string);
 void text(string,short,short,WORD);
-void num(int num,int X, int Y,WORD color);
+void num(int num, short X, short Y,WORD color);
 void clear();
 Armor* loadArmor(string);
 Weapon* loadWeapon(string);
