+++ world.h
@@ -6,7 +6,7 @@
 #include "CFmod.h"
 #include "Fmod.h"
 
-void num(int num,int X, int Y,WORD color);
+void num(int num, short X, short Y,WORD color);
 void text(string, short, short, WORD);
 void clear();
 void clrbottom();
